requirements.txt
# Core deep learning framework
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# Scientific computing & data processing
numpy==1.24.3
scipy==1.13.1
scikit-learn==1.6.1
matplotlib==3.9.4
pillow==11.3.0

# Experiment utilities
PyYAML==6.0.3
tqdm
requests==2.32.5